move GenesisState to store
