+++ utilities/genesis/Cargo.toml
@@ -13,6 +13,6 @@ description = "Generate a JSON file representing the chain state at genesis"
 [dependencies]
 anyhow = { version = "1" }
 clap = { version = "4.4", features = ["derive"] }
-miden_objects = { workspace = true, default-features = true, features = ["serde"] }
+miden-node-store = { path = "../../store" }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0" }
